+++ Tests/TestTrackInfo.cs
@@ -1,5 +1,5 @@
 using Godot;
-using XanaduProject.Data;
+using XanaduProject.DataStructure;
 
 namespace XanaduProject.Tests;
